Move formatting functions to helper file
