+++ routes/html-routes.js
@@ -5,6 +5,7 @@ const {
   searchGames,
 } = require('./helper/gamespot.js')
 const db = require('../models')
+const { formatOwnedGames, markOwned } = require('./helper/ownedGames.js')
 
 router.get('/', async function (req, res) {
   try {
@@ -45,26 +46,5 @@ router.get('/search/:game', async function (req, res) {
     res.status(401).json({ error: err })
   }
 })
-
-function formatOwnedGames(ownedGames) {
-  return ownedGames.map((game) => {
-    game.dataValues.Images = game.dataValues.Images.map(
-      (image) => image.dataValues
-    )
-    return game.dataValues
-  })
-}
-
-function markOwned(currentGames, ownedGames) {
-  for (const game of currentGames) {
-    const foundGame = ownedGames.find((data) => game.apiId === data.apiId)
-    if (foundGame) {
-      game.owned = true
-      game.id = foundGame.id
-    } else {
-      game.owned = false
-    }
-  }
-}
 
 module.exports = router
